generate_key compatibility with OTP 20.0
The old implementation was tested against OTP 20.1, which had slightly
different internal library versions than OTP 20.0, which introduced the
generate_key feature.

I've lowered the version requirement for Erlang public_key from 1.5 to
1.4.1 and switched to Version.match?() to do proper version checking.

The internal data format for RSAPrivateKey was not consistently
generated in OTP 20.0, generating version=0, but when parsing the PEM
version it returned version=:"two-prime". Turns out this is an internal
code vs named code thing, and they are the same.