Fixes Base64 encoding of generated AES keys
Prior to this change, if you selected the `:base64` format when calling
`ExCrypto.generate_aes_key/2` you would get back a string that was
encoded using `Base.url_encode64`. However, the url-encoding allows
certain characters to be included that are not strictly base 64
characters (e.g. "-" and "_"), which prevents it from working properly
when used in a context that requires strict base 64 encoding. Since the
format is called `:base64` and not something like `:base64_url`, this is
surprising behavior.

(In particular, I was trying to send the value to
[Vault](https://www.vaultproject.io) and was randomly getting back 400
errors that turned out to be happening when the generated key happened
to contain the invalid characters.)